requirements.txt
requests==2.31.0
Pillow==10.2.0
supabase==2.3.1
python-dotenv==1.0.1
torch==2.2.1
transformers==4.38.2
numpy==1.24.3
sentencepiece==0.2.0 